+++ packages/ttml-adapter/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @sub37/ttml-adapter
 
+## **1.0.1** (06 Jun 2026)
+
+- Added missing support to encoded and predefined entities decoding (#23)
+- Fixed special semantics support when default region applies (#25)
+
+---
+
 ## **1.0.0** (19 Jun 2026)
 
 - First version released